\name{throwError}
\alias{throwError}
\title{pretty prints error messages...}
\usage{throwError(apiResult)}
\description{pretty prints error messages}
\details{This function pretty prints error messages returned by the API}
\value{the error message returned by the API in human readable form}
\note{this function is not to be called directly by the user}
\references{\url{http://www.mediawiki.org/wiki/API:Errors_and_warnings}}
\author{Peter Konings \email{peter.konings@esat.kuleuven.be}}
\arguments{\item{apiResult}{a character string in XML format returned by the MediaWiki API}}
