\name{throwFailure}
\alias{throwFailure}
\title{returns failures...}
\usage{throwFailure(apiResult)}
\description{returns failures}
\details{This function returns failure messages returned by the API}
\value{the failure message returned by the API}
\note{this function is not to be called directly by the user}
\references{\url{http://www.mediawiki.org/wiki/API:Errors_and_warnings}}
\author{Peter Konings \email{peter.konings@esat.kuleuven.be}}
\arguments{\item{apiResult}{a character string in XML format returned by the MediaWiki API}}
