\name{wikirobot-package}
\alias{wikirobot-package}
\alias{wikirobot}
\docType{package}
\title{wikiRobot provides an R-MediaWiki interface and a Sweave driver for MediaWiki markup
}
\description{wikiRobot provides access to the MediaWiki API and a Sweave driver for MediaWiki markup, enabling collaborative reporting using Mediawiki
}
\details{
\tabular{ll}{
Package: \tab wikirobot\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2009-12-23\cr
License: \tab GPL (>=3)\cr
LazyLoad: \tab yes\cr
}

}
\author{
Peter Konings

Maintainer: Peter Konings <peter.konings@esat.kuleuven.be>
}
\references{
http://wikirobot.r-forge.r-project.org
http://www.mediawiki.org/API
http://www.stat.uni-muenchen.de/~leisch/Sweave/
}

\keyword{ package }
